refactor: 공통 예외 코드 디렉토리 변경
